Extract the pure SFTP connect and host-key helpers into sftpConnect
Second incremental fileSyncConnection.ts extraction, pure code motion: hostKeyBlob, settleVerify, the SFTP provider-option and algorithms allowlist constants, and the PresentedHostKey type move to packages/core/src/connection/sftpConnect.ts, with PresentedHostKey re-exported so the public surface is unchanged. The connect-option builders and host verifiers stay behind for now; they close over the connection's log, role, and client state, and their disentanglement belongs to the class split.
